requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==3.0.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.7
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
flake8==7.0.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
gspread==5.12.4
idna==3.6
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
ngrok==0.12.1
oauthlib==3.2.2
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.1.14
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
